+++ packages/networks/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lit-protocol/networks
 
+## 8.0.0-alpha.26
+
+### Major Changes
+
+- add missing `withdraw` ledge abi method
+
+## 8.0.0-alpha.25
+
+### Major Changes
+
+- add payment delegation manager
+
+## 8.0.0-alpha.24
+
+### Major Changes
+
+- add payment delegation apis
+
 ## 8.0.0-alpha.23
 
 ### Major Changes